EUREKA-810: Finish integration tests

+++ src/folio_upm/integration/services/role_service.py
@@ -53,13 +53,17 @@ class RoleService(metaclass=SingletonMeta):
         role_ids_to_delete = [hash_role.role.id for hash_role in cleanup_records if self.__should_delete(hash_role)]
         role_ids_to_delete = [rid for rid in role_ids_to_delete if rid is not None]
         total_roles = len(role_ids_to_delete)
+        if not role_ids_to_delete:
+            self._log.info("No roles to remove.")
+            return []
+
         self._log.debug("Removing roles %s: %s ...", total_roles, role_ids_to_delete)
-        roles_counter = 1
+        roles_counter = 0
         remove_rs = list[HttpRequestResult]()
         for role_id in role_ids_to_delete:
             remove_rs.append(self.__delete_role_safe(role_id))
-            self._log.info("Roles remove processed: %s/%s", roles_counter, total_roles)
             roles_counter += 1
+            self._log.info("Roles remove processed: %s/%s", roles_counter, total_roles)
         self._log.info("Roles removed successfully %s: %s", roles_counter, role_ids_to_delete)
         return remove_rs
